# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Recyclic API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth",
          "auth"
        ],
        "summary": "Login",
        "description": "Authentifie un utilisateur via son nom d'utilisateur et mot de passe, et retourne un JWT.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/signup": {
      "post": {
        "tags": [
          "auth",
          "auth"
        ],
        "summary": "Signup",
        "description": "Crée un nouveau compte utilisateur en attente de validation.",
        "operationId": "signup_api_v1_auth_signup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignupResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/forgot-password": {
      "post": {
        "tags": [
          "auth",
          "auth"
        ],
        "summary": "Forgot Password",
        "description": "Génère un token de réinitialisation et l'envoie par email (ou le log).",
        "operationId": "forgot_password_api_v1_auth_forgot_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForgotPasswordResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/reset-password": {
      "post": {
        "tags": [
          "auth",
          "auth"
        ],
        "summary": "Reset Password",
        "description": "Réinitialise le mot de passe avec un token valide.",
        "operationId": "reset_password_api_v1_auth_reset_password_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResetPasswordResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/settings/settings/alert-thresholds": {
      "get": {
        "tags": [
          "admin",
          "settings",
          "admin",
          "settings"
        ],
        "summary": "Get Alert Thresholds",
        "operationId": "get_alert_thresholds_api_v1_admin_settings_settings_alert_thresholds_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "site_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Site Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertThresholdsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "admin",
          "settings",
          "admin",
          "settings"
        ],
        "summary": "Put Alert Thresholds",
        "operationId": "put_alert_thresholds_api_v1_admin_settings_settings_alert_thresholds_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlertThresholdsUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertThresholdsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AlertThresholds-Input": {
        "properties": {
          "cash_discrepancy": {
            "type": "number",
            "minimum": 0.0,
            "title": "Cash Discrepancy",
            "description": "Cash discrepancy avant alerte"
          },
          "low_inventory": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Low Inventory",
            "description": "Nombre d'articles minimum avant alerte d'inventaire"
          }
        },
        "type": "object",
        "required": [
          "cash_discrepancy",
          "low_inventory"
        ],
        "title": "AlertThresholds",
        "description": "Dashboard alert thresholds persisted server-side."
      },
      "AlertThresholds-Output": {
        "properties": {
          "cashDiscrepancy": {
            "type": "number",
            "minimum": 0.0,
            "title": "Cashdiscrepancy",
            "description": "Cash discrepancy avant alerte"
          },
          "lowInventory": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Lowinventory",
            "description": "Nombre d'articles minimum avant alerte d'inventaire"
          }
        },
        "type": "object",
        "required": [
          "cashDiscrepancy",
          "lowInventory"
        ],
        "title": "AlertThresholds",
        "description": "Dashboard alert thresholds persisted server-side."
      },
      "AlertThresholdsResponse": {
        "properties": {
          "thresholds": {
            "$ref": "#/components/schemas/AlertThresholds-Output"
          },
          "siteId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Siteid",
            "description": "Site concerne par cette configuration"
          }
        },
        "type": "object",
        "required": [
          "thresholds"
        ],
        "title": "AlertThresholdsResponse",
        "description": "Response payload for threshold retrieval."
      },
      "AlertThresholdsUpdate": {
        "properties": {
          "thresholds": {
            "$ref": "#/components/schemas/AlertThresholds-Input"
          },
          "site_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site Id",
            "description": "Site cible pour la configuration"
          }
        },
        "type": "object",
        "required": [
          "thresholds"
        ],
        "title": "AlertThresholdsUpdate",
        "description": "Request payload for updating alert thresholds."
      },
      "AuthUser": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "telegram_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Telegram Id"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "first_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Name"
          },
          "last_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "role",
          "is_active"
        ],
        "title": "AuthUser",
        "description": "Représentation de l'utilisateur retournée après authentification."
      },
...
      "ForgotPasswordRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Adresse email de l'utilisateur"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "ForgotPasswordRequest",
        "description": "Schéma de requête pour demander la réinitialisation du mot de passe."
      },
      "ForgotPasswordResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Message de confirmation"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ForgotPasswordResponse",
        "description": "Réponse pour la demande de réinitialisation du mot de passe."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username",
            "description": "Nom d'utilisateur"
          },
          "password": {
            "type": "string",
            "title": "Password",
            "description": "Mot de passe"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest",
        "description": "Schéma de requête pour la connexion via nom d'utilisateur et mot de passe."
      },
      "LoginResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token",
            "description": "JWT d'accès"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "description": "Type de token",
            "default": "bearer"
          },
          "user": {
            "$ref": "#/components/schemas/AuthUser"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "user"
        ],
        "title": "LoginResponse",
        "description": "Réponse d'authentification contenant le token et l'utilisateur."
      },
...
      "ResetPasswordRequest": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token",
            "description": "Token de réinitialisation"
          },
          "new_password": {
            "type": "string",
            "title": "New Password",
            "description": "Nouveau mot de passe"
          }
        },
        "type": "object",
        "required": [
          "token",
          "new_password"
        ],
        "title": "ResetPasswordRequest",
        "description": "Schéma de requête pour réinitialiser le mot de passe."
      },
      "ResetPasswordResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Message de confirmation"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ResetPasswordResponse",
        "description": "Réponse pour la réinitialisation du mot de passe."
      },
...
      "SignupRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 50,
            "minLength": 3,
            "title": "Username",
            "description": "Nom d'utilisateur"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "title": "Password",
            "description": "Mot de passe"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email",
            "description": "Adresse email (optionnelle)"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "SignupRequest",
        "description": "Schéma de requête pour l'inscription d'un nouvel utilisateur."
      },
      "SignupResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Message de confirmation"
          },
          "user_id": {
            "type": "string",
            "title": "User Id",
            "description": "ID de l'utilisateur créé"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Statut du compte"
          }
        },
        "type": "object",
        "required": [
          "message",
          "user_id",
          "status"
        ],
        "title": "SignupResponse",
        "description": "Réponse d'inscription d'un nouvel utilisateur."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}